backend/tex_backend/requirements.txt
pandas==1.3.5
# pyodbc==4.0.32
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-environ==0.8.1
django-pandas==0.6.6
django-after-response==0.2.2
django-model-utils==1.1.0
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
newspaper3k==0.2.8
yfinance==0.1.63
bs4==0.0.1
pandas-datareader==0.10.0
# CurrencyConverter==0.16.5
# forex-python==1.8
py-currency-converter==1.2.0
pandas-market-calendars==3.2
full-fred==0.0.9a3
PyMySQL==1.0.2
# lxml==4.6.3
openpyxl==3.0.10
html5lib==1.1
pyxlsb==1.0.9
whitenoise==6.3.0